Move helper class `InMemoryFeedStore` into its own file
