+++ EssentialApp/EssentialAppTests/Helpers/InMemoryFeedStore.swift
@@ -1,0 +1,53 @@
+//
+//  InMemoryFeedStore.swift
+//  EssentialAppTests
+//
+//  Created by Jair Moreno Gaspar on 17/02/21.
+//
+
+import EssentialFeed
+
+class InMemoryFeedStore: FeedStore, FeedImageDataStore {
+    
+    private(set) var feedCache: CacheFeed?
+    private var feedImageDataCache: [URL: Data] = [:]
+    
+    private init(feedCache: CacheFeed? = nil) {
+        self.feedCache = feedCache
+    }
+    
+    func deleteCachedFeed(completion: @escaping DeletionCompletion) {
+        feedCache = nil
+        completion(.success(()))
+    }
+    
+    func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
+        feedCache = CacheFeed(feed: feed, timestamp: timestamp)
+        completion(.success(()))
+    }
+    
+    func retrieve(completion: @escaping RetrievalCompletion) {
+        completion(.success(feedCache))
+    }
+    
+    func insert(_ data: Data, for url: URL, completion: @escaping (InsertionResult) -> Void) {
+        feedImageDataCache[url] = data
+        completion(.success(()))
+    }
+    
+    func retrieve(dataForURL url: URL, completion: @escaping (FeedImageDataStore.RetrievalResult) -> Void) {
+        completion(.success(feedImageDataCache[url]))
+    }
+    
+    static var empty: InMemoryFeedStore {
+        InMemoryFeedStore()
+    }
+
+    static var withExpiredFeedCache: InMemoryFeedStore {
+        InMemoryFeedStore(feedCache: CacheFeed(feed: [], timestamp: Date.distantPast))
+    }
+    
+    static var withNonExpiredFeedCache: InMemoryFeedStore {
+        InMemoryFeedStore(feedCache: CacheFeed(feed: [], timestamp: Date()))
+    }
+}
